# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "",   
    "version": "1.0.2",
    "title": "testdb",
    "termsOfService": "",
    "contact": {
      "email": "shubhamjulaha.rexweb@gmail.com"
    },
    "license": {
      "name": "",
      "url": ""
    }
  },
  "host": "192.168.43.50:3000",
  "basePath": "",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/department": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "Department"
        ],
        "summary": "This API will provide you the list of all departments with all the fields",
        "description": "List of all departments",
        "operationId": "getDepartment",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "400": {
            "description": "Please FIll all fields correctly"
          },
          "200": {
            "description": "Will send `Authenticated`"
          }
        }
      },
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "Department"
        ],
        "summary": " This API is for adding a new department into the database",
        "description": "Add a new department into db",
        "operationId": "addDepartment",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": [
              "departmentName"
            ],
            "schema": {
              "$ref": "#/definitions/Department"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Please fill all fields correctly"
          },
          "200": {
            "description": "Success with data"
          }
        }
      }
    },
...
    "/api/PerformanceMatrixMainParameters": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "Performance Matrix Main Parameters"
        ],
        "summary": "This API will provide you the list of all main parameters of PerformanceMatrix",
        "description": "List of all Performance Matrix",
        "operationId": "getPerformance Matrix",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "400": {
            "description": "Please FIll all fields correctly"
          },
          "200": {
            "description": "Success with data"
          }
        }
      },
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "Performance Matrix Main Parameters"
        ],
        "summary": " This API will add a new main category in performance matrix module ",
        "description": "Add a new performance matrix into db",
        "operationId": "addPerformance",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Performance that needs to be added to the store",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Performance"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Please fill all fields correctly"
          },
          "200": {
            "description": "Success with data"
          }
        }
      }
    },
...
    "/api/PerformanceMatrixSubParameters": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "Performance Matrix Sub Parameters"
        ],
        "summary": "This API will provide us the list of all sub categories of performnace matrix module",
        "description": "List of all sub Performance Matrix",
        "operationId": "getSubPerformance Matrix",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "400": {
            "description": "Please FIll all fields correctly"
          },
          "200": {
            "description": "Success with data"
          }
        }
      },
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "Performance Matrix Sub Parameters"
        ],
        "summary": " This API is for adding new sub category to performance matrix sub parameters ",
        "description": "Add a new sub-performance matrix into db",
        "operationId": "addPerformance",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Performance that needs to be added to the store",
            "required": false,
            "schema": {
              "$ref": "#/definitions/subPerformance"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Please fill all fields correctly"
          },
          "200": {
            "description": "Success with data"
          }
        }
      }
    },
...
    "/api/matrixCalculation/addcalculation": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "Performance Matrix Calculation"
        ],
        "summary": "This API will provide listing of all parameters of  MatrixParameters/SubMatrixParameters/Evaluation",
        "description": "List of all matrix calculations",
        "operationId": "getDepartment",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "400": {
            "description": "Please FIll all fields correctly"
          },
          "200": {
            "description": "Success with data"
          }
        }
      }
    },
...
    "/api/getPermissions": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "Permissions"
        ],
        "summary": "This API will provide listing of all parameters of  MatrixParameters/SubMatrixParameters/Evaluation",
        "description": "List of all matrix calculations",
        "operationId": "getDepartment",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "400": {
            "description": "Please FIll all fields correctly"
          },
          "200": {
            "description": "Success with data"
          }
        }
      }
    },
...
    "/api/CRM": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "CRM Module"
        ],
        "summary": "This API will return list of all clients . ",
        "description": "This module will return list of all clients along with personal information",
        "operationId": "getCRMfields",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "400": {
            "description": "Please FIll all fields correctly"
          },
          "200": {
            "description": "Success with data"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
...
    "Department": {
      "type": "object",
      "properties": {
        "departmentName": {
          "type": "string"
        },
        "shortCode": {
          "type": "string"
        },
        "kra": {
          "type": "string"
        },
        "isActive": {
          "type": "integer"
        }
      }
    },
...
    "Performance": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "percentage": {
          "type": "integer"
        },
        "isActive": {
          "type": "integer"
        }
      }
    },
    "subPerformance": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "parentId": {
          "type": "integer"
        },
        "isActive": {
          "type": "integer"
        }
      }
    },
...
  },
...
}